First commit with data
The data provided here is made from the catalog of Oleg

 @book{Oleg_Hodjash_1998, place={Fribourg},
title={Catalogue of the Monuments of Ancient Egypt: From the Museums of the Russian Federation, Ukraine, Bielorussia, Caucasus, Middle Asia and the Baltic States}, url={https://doi.org/10.5167/uzh-159792}, author={Oleg, Berlev and Hodjash, Svetlana}, year={1998} }

I have included pleiades ids whereever possible. They need to be dissociated from
the description to be included in a seperate field in json if possible.